refactor(driver): extract standalone types/consts/helpers to driver/types.rs (refactor 2.1e)
Fifth commit of the Phase 2.1 mechanical split. Moves all remaining
standalone items (types, constants, helper functions) out of
driver/mod.rs into a dedicated sibling module. After this commit,
`driver/mod.rs` contains **only** the top-level `use` statements, the
`Driver<P, Pool>` struct, the giant `impl Driver` block, and the test
mod — everything else lives in a sibling.

- `crates/based-rollup/src/driver/types.rs` (new, 279 lines):
  - **Types**: `DriverMode`, `VerificationDecision`,
    `TriggerExecutionResult`, `BuiltBlock`, `L1ConfirmedAnchor`,
    `TxJournalEntry` + impl.
  - **Constants**: `TX_JOURNAL_STAGE_ID`, `FORK_CHOICE_DEPTH`,
    `CHECKPOINT_INTERVAL`, `MAX_BACKOFF_SECS`,
    `SUBMISSION_COOLDOWN_SECS`, `MAX_BATCH_SIZE`,
    `MAX_PENDING_SUBMISSIONS`, `MAX_PENDING_CROSS_CHAIN_ENTRIES`,
    `MAX_CONSECUTIVE_FAILURES`, `MIN_L1_CALL_INTERVAL`,
    `FCU_SYNCING_MAX_RETRIES`, `FCU_SYNCING_INITIAL_BACKOFF_MS`,
    `DESIRED_GAS_LIMIT`.
  - **Helper functions**: `encode_block_transactions`,
    `calc_gas_limit`, `compute_forkchoice_state`.
  - Visibility: internal items are `pub(super)` so the `impl Driver`
    block in mod.rs can still use them by private name; public API
    types (`DriverMode`, `BuiltBlock`) are `pub` and re-exported.

- `crates/based-rollup/src/driver/mod.rs` (-240 lines → 4584 total):
  - `mod types;` declaration.
  - `pub use types::{BuiltBlock, DriverMode};` — preserves the public
    API.
  - `use types::{...};` private use list brings all constants,
    private types, and helper fns into scope for the `impl Driver`
    block without touching any internal code.
  - `use reth_stages_types::StageCheckpoint;` — narrowed from
    `{StageCheckpoint, StageId}` (StageId is no longer referenced
    directly in mod.rs; the `TX_JOURNAL_STAGE_ID` const in types.rs
    owns it).

Cumulative progress:
- mod.rs: 5509 → 4584 lines (-925 / -16.8%)
- 4 sibling modules: hold (199) + pending_queue (202) +
  flush_plan (321) + types (279) = 1001 lines of well-bounded code
  that used to live in the same 5500-line wall.

`cargo check`, `cargo clippy --all-features -- -D warnings`, and
`cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
change.

Next: 2.1f starts extracting impl blocks. Candidates (in decreasing
blast radius order): verify.rs (verification + filtering), journal.rs
(tx journal methods), rewind.rs (rewind methods), protocol_txs.rs
(build_builder_protocol_txs), build.rs (build_and_insert_block +
forkchoice), flush.rs (flush_to_l1 + helpers), step_builder.rs
(step_builder + drainers).
